# frontend/openapi.json
{
...
  "info": {
    "description": "Canonical Django REST API used by Moksha clients.",
    "title": "Moksha AI API",
    "version": "1.0.0"
  },
  "openapi": "3.0.2",
  "paths": {
...
    "/api/v1/chats/{id}/": {
      "delete": {
        "description": "Delete a chat session.",
        "operationId": "deleteApiV1ChatsById",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": ["api"]
      },
      "get": {
        "description": "Get a chat with its full message history.",
        "operationId": "getApiV1ChatsById",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": ""
          }
        },
        "tags": ["api"]
      }
    },
...
    "/api/v1/chats/{id}/runs/": {
      "post": {
        "description": "Create a durable generation run.",
        "operationId": "postApiV1ChatsByIdRuns",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            },
            "application/x-www-form-urlencoded": {
              "schema": {}
            },
            "multipart/form-data": {
              "schema": {}
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": ""
          }
        },
        "tags": ["api"]
      }
    },
...
    "/api/v1/runs/{id}/": {
      "get": {
        "description": "Durable generation run endpoints.",
        "operationId": "getApiV1RunsById",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {}
              }
            },
            "description": ""
          }
        },
        "tags": ["api"]
      }
    },
...
  }
}